blog/requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.7
Django==4.2.4
django-ckeditor==6.7.0
django-environ==0.10.0
django-js-asset==2.1.0
django-taggit==4.0.0
flake8==6.1.0
isort==5.12.0
mccabe==0.7.0
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
